requirements.txt
Jinja2
PyYAML
Pillow